完成會員登入資料庫串接與登入功能，新增AuthContext
