+++ app/api/member/route.js
@@ -1,92 +1,66 @@
+// app/api/member/route.js
 import { NextResponse } from 'next/server';
-import { pool } from '@/app/lib/db.ts'
+import { database } from '@/app/api/_lib/db'; // 引入你的資料庫連接模組
 import jwt from 'jsonwebtoken';
 
-console.log('JWT_SECRET:', process.env.JWT_SECRET);
-
-export async function GET(request) {
-  // ... GET 函數保持不變 ...
-}
-
 export async function POST(request) {
-  console.log('Received POST request');
   try {
-    // 會員登入
     const { email, password } = await request.json();
-    console.log('Received login attempt for email:', email);
 
-    // 驗證必要欄位
+    // 驗證輸入
     if (!email || !password) {
-      console.log('Missing email or password');
       return NextResponse.json(
         { success: false, message: '請輸入電子郵件和密碼' },
         { status: 400 }
       );
     }
 
-    const connection = await pool.getConnection(); // 獲取資料庫連接
-    try {
-      // 查詢用戶
-      const [users] = await connection.execute(
-        'SELECT * FROM `users` WHERE `user_email` = ? LIMIT 1',
-        [email]
+    // 查找用戶
+    const [users, queryError] = await database.executeSecureQuery(
+      'SELECT * FROM users WHERE user_email = ? LIMIT 1',
+      [email]
+    );
+
+    if (queryError || !users || users.length === 0) {
+      return NextResponse.json(
+        { success: false, message: '電子郵件或密碼錯誤' },
+        { status: 401 }
       );
-      console.log('Query result:', users);
+    }
 
-      if (users.length === 0) {
-        return NextResponse.json(
-          { success: false, message: '電子郵件或密碼錯誤' },
-          { status: 401 }
-        );
-      }
+    const user = users[0];
 
-      const user = users[0];
-
-      // 驗證密碼（明碼比對）
-      if (password !== user.user_password) {
-        return NextResponse.json(
-          { success: false, message: '電子郵件或密碼錯誤' },
-          { status: 401 }
-        );
-      }
-
-      // 生成 JWT token
+    // 比較明碼密碼 (請在生產環境中使用 bcrypt 或 Argon2 進行雜湊)
+    if (password === user.user_password) {
+      // 登入成功
       const token = jwt.sign(
         { userId: user.user_id, email: user.user_email },
         process.env.JWT_SECRET,
         { expiresIn: '1d' }
       );
 
-      // 更新最後登入時間
-      await connection.execute(
-        'UPDATE `pet_proj`.`users` SET `last_login` = NOW() WHERE `user_id` = ?',
-        [user.user_id]
-      );
-
-      // 返回成功響應
       return NextResponse.json({
         success: true,
         message: '登入成功',
         data: {
           token,
           user: {
-            user_id: user.user_id,
-            user_name: user.user_name,
-            user_email: user.user_email,
+            id: user.user_id,
+            name: user.user_name,
+            email: user.user_email,
           },
         },
       });
-    } finally {
-      connection.release(); // 釋放資料庫連接
+    } else {
+      return NextResponse.json(
+        { success: false, message: '電子郵件或密碼錯誤' },
+        { status: 401 }
+      );
     }
   } catch (error) {
     console.error('登入時發生錯誤:', error);
     return NextResponse.json(
-      { 
-        success: false, 
-        message: '登入時發生錯誤，請稍後再試', 
-        error: error.message
-      },
+      { success: false, message: '登入時發生錯誤，請稍後再試' },
       { status: 500 }
     );
   }
