+++ app/layout.js
@@ -3,6 +3,7 @@ import 'bootstrap/dist/css/bootstrap.min.css'
 import './globals.css'
 import Script from 'next/script'
 import LayoutWrapper from './LayoutWrapper'
+import { AuthProvider } from './context/AuthContext';
 
 const fontIansui = localFont({
   src: '../public/fonts/Iansui-Regular.ttf',
@@ -19,7 +20,9 @@ export const metadata = {
     '這是一個寵物領養網站，旨在提供寵物領養服務，讓更多毛孩能有溫暖的家。',
 }
 
+// 客戶端組件已移至 LayoutWrapper.js
 export default function RootLayout({ children }) {
+  console.log('AuthProvider rendered'); // 添加此行
   return (
     <html lang="zh-hant-tw" suppressHydrationWarning>
       <body
@@ -31,7 +34,9 @@ export default function RootLayout({ children }) {
         }}
         suppressHydrationWarning
       >
-        <LayoutWrapper>{children}</LayoutWrapper>
+        <AuthProvider>
+          <LayoutWrapper>{children}</LayoutWrapper>
+        </AuthProvider>
 
         <Script
           src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"
@@ -39,6 +44,5 @@ export default function RootLayout({ children }) {
         />
       </body>
     </html>
-  )
+  );
 }
-// 客戶端組件已移至 LayoutWrapper.js
