+++ app/member/page.js
@@ -1,167 +1,108 @@
-'use client';
-import React, { useState, useEffect } from 'react';
-import styles from './login.module.css';
-import { useRouter } from 'next/navigation';
-import Link from 'next/link';
-import Swal from 'sweetalert2';
+'use client'
+
+import React, { useState, useEffect } from 'react'
+import Link from 'next/link'
+import styles from "./member.module.css";
 
 export default function MemberPage() {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [rememberMe, setRememberMe] = useState(false);
-  const router = useRouter();
-
+  const [userData, setUserData] = useState(null);
+  
   useEffect(() => {
-    const rememberedEmail = localStorage.getItem('rememberedEmail');
-    if (rememberedEmail) {
-      setEmail(rememberedEmail);
-      setRememberMe(true);
-    }
+    // 從API獲取用戶數據
+    const fetchUserData = async () => {
+      try {
+        // 從 localStorage 獲取 token
+        const token = localStorage.getItem('token');
+        
+        if (!token) {
+          console.error('No token found, user might not be logged in');
+          return;
+        }
+
+        const response = await fetch('/api/member', {
+          headers: {
+            'Authorization': `Bearer ${token}`
+          }
+        });
+
+        if (!response.ok) {
+          throw new Error('Failed to fetch user data');
+        }
+
+        const data = await response.json();
+        setUserData(data);
+      } catch (error) {
+        console.error('Error fetching user data:', error);
+      }
+    };
+
+    fetchUserData();
   }, []);
 
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (!email || !password) {
-      await Swal.fire({
-        title: '錯誤',
-        text: '請填寫所有欄位',
-        icon: 'error',
-        confirmButtonText: '確定'
-      });
-      return;
-    }
-
+  const handleNicknameChange = async (event) => {
+    const newNickname = event.target.value;
     try {
-      const response = await fetch('/api/member', {
+      const token = localStorage.getItem('token');
+      const response = await fetch('/api/user/updateNickname', {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
+          'Authorization': `Bearer ${token}`
         },
-        body: JSON.stringify({ email, password, rememberMe }),
+        body: JSON.stringify({ nickname: newNickname }),
       });
-
-      const data = await response.json();
-
-      if (response.ok && data.success) {
-        localStorage.setItem('token', data.data.token);
-        localStorage.setItem('user', JSON.stringify(data.data.user));
-        if (rememberMe) {
-          localStorage.setItem('rememberedEmail', email);
-        } else {
-          localStorage.removeItem('rememberedEmail');
-        }
-        
-        await Swal.fire({
-          title: '登入成功！',
-          icon: 'success',
-          timer: 1500,
-          showConfirmButton: false
-        });
-        
-        router.push('/member/MemberCenter');
+      if (response.ok) {
+        setUserData(prevData => ({ ...prevData, nickname: newNickname }));
       } else {
-        await Swal.fire({
-          title: '登入失敗',
-          text: data.message || '請檢查您的電子郵件和密碼。',
-          icon: 'error',
-          confirmButtonText: '確定'
-        });
+        throw new Error('Failed to update nickname');
       }
     } catch (error) {
-      console.error('登入請求失敗:', error);
-      await Swal.fire({
-        title: '錯誤',
-        text: '登入時發生錯誤，請稍後再試。',
-        icon: 'error',
-        confirmButtonText: '確定'
-      });
+      console.error('Error updating nickname:', error);
     }
   };
 
   return (
     <>
-      <header className={styles.headerSection}>
-        <img
-          src="https://cdn.builder.io/api/v1/image/assets/TEMP/c961159506ebe222e2217510289e3eee7203e02a0affe719332fe812045a0061?placeholderIfAbsent=true&apiKey=2d1f7455128543bfa30579a9cce96321"
-          alt="Header background"
-          className={styles.headerBackground}
-        />
-        <h1 className={styles.pageTitle}>會員登入</h1>
-      </header>
-
-      <div className={styles.formContainer}>
-        <h2 className={styles.sectionTitle}>登入會員</h2>
-        <form onSubmit={handleSubmit} className={styles.form}>
-          <div className={styles.formGroup}>
-            <label htmlFor="email" className={styles.formLabel}>
-              電子信箱 :
-            </label>
-            <input
-              type="email"
-              id="email"
-              className={styles.formInput}
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              required
-            />
-
-            <label htmlFor="password" className={styles.formLabel}>
-              密碼 :
-            </label>
-            <input
-              type="password"
-              id="password"
-              className={styles.formInput}
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              required
-            />
-
-            <div className={styles.rememberMe}>
-              <input
-                type="checkbox"
-                id="rememberMe"
-                name="rememberMe"
-                checked={rememberMe}
-                onChange={(e) => setRememberMe(e.target.checked)}
-              />
-              <label htmlFor="rememberMe">記住我</label>
+      <main className={styles.profile_page}>
+        <div className={styles.member_container}>
+          <section className={styles.profile_section}>
+            <div className={styles.profile_photos}>
+              <img src="https://cdn.builder.io/api/v1/image/assets/TEMP/c07e5bb4325caeb94efd091416f6964123f3611a" alt="大頭照" className="profile-photo" />
+              <select value={userData?.nickname || ""} onChange={handleNicknameChange}>
+                <option value="">暱稱</option>
+                <option value="愛心小天使">愛心小天使</option>
+                <option value="乾爹乾媽">乾爹乾媽</option>
+                <option value="拔鼻媽咪">拔鼻媽咪</option>
+                <option value="超級拔鼻媽咪">超級拔鼻媽咪</option>
+              </select>
             </div>
-
-            <button
-              type="submit"
-              className="button"
-              style={{ width: '200px', height: '50px', fontSize: '28px' }}
-            >
-              登入
-            </button>
-          </div>
-        </form>
-
-        <div>
-          <p className={styles.loginLink}>
-            還沒有會員?  
-            <Link
-              href="/member/MemberLogin/register"
-              className={styles.link}
-              style={{ fontSize: '22px' }}
-            >
-              點此註冊
-            </Link>
-          </p>
-          <p className={styles.loginLink}>
-            忘記密碼?  
-            <Link
-              href="/member/MemberLogin/forgot"
-              className={styles.link}
-              style={{ fontSize: '22px' }}
-            >
-              請點我
-            </Link>
-          </p>
+                      
+            <div className={styles.profile_content}>
+              <h3 className={styles.profile_title}>個人資料</h3>
+              <div className={styles.profile_group}>
+                <label className={styles.form_label}>姓名 : {userData?.user_name}</label>
+                <hr />
+                <label className={styles.form_label}>電話 : {userData?.user_phone}</label>
+                <hr />
+                <label className={styles.form_label}>生日 : {userData?.user_birthday}</label>
+                <hr />
+                <label className={styles.form_label}>論壇ID : {userData?.user_id}</label>
+                <hr />
+                <label className={styles.form_label}>地址 : {userData?.user_address}</label>
+                <hr />
+              </div>
+              
+              <div className={styles.edit_but}>
+                <button
+                  className="button"
+                  style={{ width: '150px', height: '40px', fontSize: '20px' }}>
+                  <Link href="/member/memsetting">修改資料</Link>
+                </button>
+              </div>
+            </div>
+          </section>
         </div>
-      </div>
+      </main>
     </>
-  );
+  )
 }
